更改窗口和显示尺寸类型为 uint
将 Display 结构体中的 mainSize 和 workSize 类型从 Vector2D<int> 更改为 Vector2D<uint>。更新 IWindowEvents 和 IWindowProperties 接口中的相关事件和属性类型。修改 Window 类的字段和属性类型，并在调用相关方法时进行类型转换。更新 WindowController 和 WindowUtils 中的方法参数和逻辑，以支持新的 uint 类型。

+++ src/ZenithEngine.Windowing/Window.cs
@@ -156,12 +156,12 @@ internal unsafe partial class Window : IWindow
         Handle = WindowUtils.Sdl.CreateWindow(Title,
                                               Position.X,
                                               Position.Y,
-                                              Size.X,
-                                              Size.Y,
+                                              (int)Size.X,
+                                              (int)Size.Y,
                                               (uint)flags);
 
-        WindowUtils.Sdl.SetWindowMinimumSize(Handle, MinimumSize.X, MinimumSize.Y);
-        WindowUtils.Sdl.SetWindowMaximumSize(Handle, MaximumSize.X, MaximumSize.Y);
+        WindowUtils.Sdl.SetWindowMinimumSize(Handle, (int)MinimumSize.X, (int)MinimumSize.Y);
+        WindowUtils.Sdl.SetWindowMaximumSize(Handle, (int)MaximumSize.X, (int)MaximumSize.Y);
         WindowUtils.Sdl.SetWindowOpacity(Handle, Opacity);
 
         WindowController.AddLoop(this);
